requirements.txt
Django==2.2.6
jsonschema==3.1.1
uWSGI==2.0.18
django-service-objects==0.5.0